requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
odfpy>=1.4.1
python-docx>=0.8.11
python-dotenv>=1.0.0
openai>=1.0.0
pymongo>=4.5.0
plotly>=5.15.0
pywin32>=306; sys_platform == "win32" 